+++ plugins/UserSettings/API.php
@@ -8,19 +8,7 @@
  */
 namespace Piwik\Plugins\UserSettings;
 
-use Piwik\Archive;
-use Piwik\DataTable;
-use Piwik\Metrics;
-use Piwik\Piwik;
-use Piwik\Plugins\DevicesDetection\Archiver AS DDArchiver;
-use Piwik\Plugins\CoreHome\Columns\Metrics\VisitsPercent;
-
-/**
- * @see plugins/UserSettings/functions.php
- */
-require_once PIWIK_INCLUDE_PATH . '/plugins/UserSettings/functions.php';
-
-/**
+/*
  * The UserSettings API lets you access reports about some of your Visitors technical settings:
  * plugins supported in their browser, Screen resolution and Screen types (normal, widescreen, dual screen or mobile).
  *
@@ -28,17 +16,6 @@ require_once PIWIK_INCLUDE_PATH . '/plugins/UserSettings/functions.php';
  */
 class API extends \Piwik\Plugin\API
 {
-    protected function getDataTable($name, $idSite, $period, $date, $segment)
-    {
-        Piwik::checkUserHasViewAccess($idSite);
-        $archive = Archive::build($idSite, $period, $date, $segment);
-        $dataTable = $archive->getDataTable($name);
-        $dataTable->filter('Sort', array(Metrics::INDEX_NB_VISITS));
-        $dataTable->queueFilter('ReplaceColumnNames');
-        $dataTable->queueFilter('ReplaceSummaryRowLabel');
-        return $dataTable;
-    }
-
     /**
      * @deprecated since 2.10.0   See {@link Piwik\Plugins\Resolution\API} for new implementation.
      */
@@ -77,7 +54,6 @@ class API extends \Piwik\Plugin\API
     }
 
     /**
-     * Gets a DataTable displaying number of visits by device type.
      * @deprecated since 2.10.0   See {@link Piwik\Plugins\DevicesDetector\API} for new implementation.
      */
     public function getMobileVsDesktop($idSite, $period, $date, $segment = false)
@@ -117,20 +93,19 @@ class API extends \Piwik\Plugin\API
         return \Piwik\Plugins\DevicePlugins\API::getInstance()->getPlugin($idSite, $period, $date, $segment);
     }
 
+    /**
+     * @deprecated since 2.11.0   See {@link Piwik\Plugins\UserLanguage\API} for new implementation.
+     */
     public function getLanguage($idSite, $period, $date, $segment = false)
     {
-        $dataTable = $this->getDataTable(Archiver::LANGUAGE_RECORD_NAME, $idSite, $period, $date, $segment);
-        $dataTable->filter('GroupBy', array('label', __NAMESPACE__ . '\groupByLangCallback'));
-        $dataTable->filter('ColumnCallbackReplace', array('label', __NAMESPACE__ . '\languageTranslate'));
-
-        return $dataTable;
+        return \Piwik\Plugins\UserLanguage\API::getInstance()->getLanguage($idSite, $period, $date, $segment);
     }
 
+    /**
+     * @deprecated since 2.11.0   See {@link Piwik\Plugins\UserLanguage\API} for new implementation.
+     */
     public function getLanguageCode($idSite, $period, $date, $segment = false)
     {
-        $dataTable = $this->getDataTable(Archiver::LANGUAGE_RECORD_NAME, $idSite, $period, $date, $segment);
-        $dataTable->filter('ColumnCallbackReplace', array('label', __NAMESPACE__ . '\languageTranslateWithCode'));
-
-        return $dataTable;
+        return \Piwik\Plugins\UserLanguage\API::getInstance()->getLanguageCode($idSite, $period, $date, $segment);
     }
 }