+++ plugins/UserSettings/UserSettings.php
@@ -23,7 +23,6 @@ class UserSettings extends \Piwik\Plugin
     public function getListHooksRegistered()
     {
         return array(
-            'Metrics.getDefaultMetricTranslations' => 'addMetricTranslations',
             'Request.getRenamedModuleAndAction'    => 'renameDeprecatedModuleAndAction',
         );
     }
@@ -55,18 +54,12 @@ class UserSettings extends \Piwik\Plugin
             $module = 'Resolution';
         }
 
+        if ($module == 'UserSettings' && ($action == 'getLanguage' || $action == 'getLanguageCode')) {
+            $module = 'UserLanguage';
+        }
+
         if ($module == 'UserSettings' && $action == 'getPlugin') {
             $module = 'DevicePlugins';
         }
     }
-
-    public function addMetricTranslations(&$translations)
-    {
-        $metrics = array(
-            'nb_visits_percentage' => Piwik::translate('General_ColumnPercentageVisits')
-        );
-
-        $translations = array_merge($translations, $metrics);
-    }
-
 }