+++ tests/PHPUnit/Unit/DeprecatedMethodsTest.php
@@ -20,7 +20,7 @@ use ReflectionClass;
  */
 class DeprecatedMethodsTest extends \PHPUnit_Framework_TestCase
 {
-    public function test_version2_0_4()
+    public function test_deprecations()
     {
         $validTill = '2015-02-25';
         $this->assertDeprecatedMethodIsRemoved('\Piwik\Period', 'factory', $validTill);
@@ -65,6 +65,8 @@ class DeprecatedMethodsTest extends \PHPUnit_Framework_TestCase
         $this->assertDeprecatedMethodIsRemoved('Piwik\Plugins\UserSettings\API', 'getResolution', $validTill);
         $this->assertDeprecatedMethodIsRemoved('Piwik\Plugins\UserSettings\API', 'getConfiguration', $validTill);
         $this->assertDeprecatedMethodIsRemoved('Piwik\Plugins\UserSettings\API', 'getPlugin', $validTill);
+        $this->assertDeprecatedMethodIsRemoved('Piwik\Plugins\UserSettings\API', 'getLanguage', $validTill);
+        $this->assertDeprecatedMethodIsRemoved('Piwik\Plugins\UserSettings\API', 'getLanguageCode', $validTill);
         $this->assertDeprecatedMethodIsRemoved('Piwik\Plugins\UserSettings\UserSettings', 'renameDeprecatedModuleAndAction', $validTill);
 
         $this->assertDeprecatedMethodIsRemovedInPiwik3('\Piwik\Menu\MenuAbstract', 'add');