+++ tests/PHPUnit/bootstrap.php
@@ -34,10 +34,7 @@ if (!defined('PIWIK_INCLUDE_SEARCH_PATH')) {
 
 require_once PIWIK_INCLUDE_PATH . '/core/bootstrap.php';
 
-require_once PIWIK_INCLUDE_PATH . '/tests/PHPUnit/DatabaseTestCase.php';
-require_once PIWIK_INCLUDE_PATH . '/tests/PHPUnit/IntegrationTestCase.php';
-require_once PIWIK_INCLUDE_PATH . '/tests/PHPUnit/BenchmarkTestCase.php';
-require_once PIWIK_INCLUDE_PATH . '/tests/PHPUnit/FakeAccess.php';
+require_once PIWIK_INCLUDE_PATH . '/libs/PiwikTracker/PiwikTracker.php';
 require_once PIWIK_INCLUDE_PATH . '/tests/PHPUnit/TestingEnvironment.php';
 
 if (getenv('PIWIK_USE_XHPROF') == 1) {
@@ -49,7 +46,7 @@ StaticContainer::setEnvironment('test');
 
 // require test fixtures
 $fixturesToLoad = array(
-    '/tests/PHPUnit/UI/Fixtures/*.php',
+    '/tests/UI/Fixtures/*.php',
     '/plugins/*/tests/Fixtures/*.php',
     '/plugins/*/Test/Fixtures/*.php',
 );
@@ -117,6 +114,7 @@ Try again.";
         // The SQL error is for Travis...
         || strpos($response, 'Table &#039;piwik_tests.option&#039; doesn&#039;t exist') !== false
         || strpos($response, 'Table &#039;piwik_tests.piwiktests_option&#039; doesn&#039;t exist') !== false
+        || strpos($response, 'Unknown database &#039;piwik_tests&#039;') !== false
     ) {
 
         return;