+++ tests/travis/travis.sh
@@ -43,7 +43,7 @@ then
         echo ""
         echo "http://builds-artifacts.piwik.org/$artifacts_folder/$TRAVIS_JOB_NUMBER/screenshot-diffs/diffviewer.html"
         echo ""
-        echo "If the new screenshots are valid, then you can copy them over to tests/PHPUnit/UI/expected-ui-screenshots/."
+        echo "If the new screenshots are valid, then you can copy them over to tests/UI/expected-ui-screenshots/."
         echo ""
 
         if [ -n "$PLUGIN_NAME" ]